Run install vim plug command
It is run after doIt function because I think .vimrc should be in place
before this plugin get loaded into .vim folder.

I use this method instead of download it before hand and sync it along
with other files is because I don't have manually update this vim plug
version if it happen to have a new version.