Change behavior for mac sleeping
I'd experienced the power loss due to some reasons. I guessed that the
reason is the disable of the hibernation mode which suck all power from
mac while it should going to sleep. This changes is made so that the mac
go to sleep after 10 minutes of inactivity. Put current state in
harddisk in case of power loss. Finally, standby first, sleep, then
hibernate after 3600 seconds.